Fix key-vault name length constraint
Azure Key Vault names are universally unique.

To make sure there are no name collisions with the new production
Key-Vault, a random_pet resource was used as a prefix. But Azure Key
Vault has a constraint for name length (8-24 chars). To fix it,
random_string is being used instead of `random_pet`. The string length is
set to 3 to accommodate long project names and deployment environment
labels.