Allow Terraform principal to remove Key Vault
- Allow Terraform's service principal to purge Azure Key Vault entries
  and delete the vault itself.
  Currently, the key vault permissions set by terraform does not allow
Delete permission for secrets. This is fixed. So terraform service
principal that can create the vault and entries can also remove the
vault.

- Add Admin AD group to manage Key Vault and entries. Added AzureAD
  provider for this reason.

TODO:
Currently, the permissions given are dangerous. Limit and restrict
to baseline complying with principles of least privilege. So basically,
Set and Update permissions for secrets management and Delete and purge
for cleanup.