Remove AD provider in favour of variable
Remove AzureAD provider in favour of a simple variable. In an attempt to
create "clean code", AzureAD provider was used to fetch the object_id of
an admin AD-group. But this proved more complicated than expected.

The AD provider was removed in this change and a variable with an empty
default is being used for fetching object ID of the admin group. This
variable was set in the variable-sets feature in Terraform cloud