Add inital ansible config management files
We need some way of managing server configurations, policy assurances,
and a way to maintain state and detect drifts in configuration.
Terraform has proved cumbersome to use for this purpose due to the
complexity involved.

Ansible is a better choice and does not require any agents and is
idempotent (to an extent). Drifts can be checked without applying
changes. Ansible is also easy to learn and understand; it use YAML
playbooks which devs and DevOps understand with practice managing CI
configurations.

Organisation of the file hierarchy is pending. Inventory files won't be
commited to the SCM and will be maintained centrally somewhere in a
private repository accessible by devs and devops.